Mensagem descrevendo as alterações

+++ src/ProdutosDTO.java
@@ -1,38 +1,18 @@
 public class ProdutosDTO {
     private int id;
     private String nome;
-    private int valor;
+    private double valor;
     private String status;
 
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public int getValor() {
-        return valor;
-    }
-
-    public void setValor(int valor) {
-        this.valor = valor;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
+    
+    public String getNome() { return nome; }
+    public void setNome(String nome) { this.nome = nome; }
+    
+    public double getValor() { return valor; }
+    public void setValor(double valor) { this.valor = valor; }
+    
+    public String getStatus() { return status; }
+    public void setStatus(String status) { this.status = status; }
 }
